% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_title.R
\name{color_title}
\alias{color_title}
\title{Convert to Colorful Title}
\usage{
color_title(title, keywords, colors)
}
\arguments{
\item{title}{Plain text title}

\item{keywords}{Word(s) in title you want to highlight}

\item{colors}{What color(s) you want to highlight the keyword(s) with}
}
\description{
This function allows you to add the HTML needed for ggtext to highlight your title.
}
\examples{
color_title(title = "This is a really cool package",
 keywords = c("really", "cool"),
 colors = c("red","blue"))
}
